move renderAdminPage back to admin.test
